requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
Babel==2.11.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-tailwind==0.5.0
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
Django==4.1.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-money==3.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.0.0
django-shortcuts==1.6
django-templated-mail==1.1.1
et-xmlfile==1.1.0
extensions==0.4
filelock==3.8.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.23.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
phonenumberslite==8.13.0
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.3
py-moneyed==2.0
pycparser==2.21
pydantic==1.9.2
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.4.0
PyPDF3==1.0.6
pypyodbc==1.3.6
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.25.11
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0